Safari-609.1.11.1/LayoutTests/fast/events/mouse-relative-position-expected.txt

span

PASS simulateElementClick(testElement, [100, 100, 200, 200]); is '100, 100, 200, 200'
PASS simulateElementClick(spanElement, [16, 10, 16, 10]); is '16, 10, 16, 10'
PASS simulateElementClick(inputElement, [40, 10, 40, 10]); is '40, 10, 40, 10'
